requirements.txt
streamlit==1.22.0
openai==0.27.2
pandas==1.5.3
requests==2.28.2
google-auth==2.18.0
google-auth-oauthlib==0.4.6
google-auth-httplib2==0.1.0
google-api-python-client==2.83.0
python-dotenv==1.0.0
